refactor: order helpers readability
