+++ packages/swap/src/helpers/orders/getQuantityWithSlippage.ts
@@ -1,0 +1,13 @@
+import {ceilDivision} from '../../utils/ceilDivision'
+
+export const getQuantityWithSlippage = (
+  quantity: bigint,
+  slippage: bigint,
+): bigint => {
+  const slippageAmount = ceilDivision(
+    BigInt(1000) * slippage * quantity,
+    BigInt(100 * 1000),
+  )
+
+  return quantity - slippageAmount
+}
