+++ packages/swap/src/helpers/orders/types.ts
@@ -1,0 +1,6 @@
+import {Balance} from '@yoroi/types'
+
+export type AmountPair = {
+  sell: Balance.Amount
+  buy: Balance.Amount
+}
